+++ packages/server/lib/src/query.dart
@@ -4,7 +4,7 @@ import 'package:engine_io_shared/exceptions.dart';
 import 'package:engine_io_shared/transports.dart';
 import 'package:universal_io/io.dart' hide Socket;
 
-import 'package:engine_io_server/src/server/server.dart';
+import 'package:engine_io_server/src/server.dart';
 
 /// Contains the parameters extracted from HTTP queries.
 class QueryParameters {