+++ packages/server/lib/src/server.dart
@@ -5,10 +5,10 @@ import 'package:engine_io_shared/packets.dart';
 import 'package:engine_io_shared/transports.dart';
 import 'package:universal_io/io.dart' hide Socket;
 
-import 'package:engine_io_server/src/server/client_manager.dart';
-import 'package:engine_io_server/src/server/configuration.dart';
-import 'package:engine_io_server/src/server/query.dart';
-import 'package:engine_io_server/src/server/socket.dart';
+import 'package:engine_io_server/src/client_manager.dart';
+import 'package:engine_io_server/src/configuration.dart';
+import 'package:engine_io_server/src/query.dart';
+import 'package:engine_io_server/src/socket.dart';
 import 'package:engine_io_server/src/transports/polling/polling.dart';
 
 /// The engine.io server.