+++ packages/server/lib/src/transports/websocket/websocket.dart
@@ -8,7 +8,7 @@ import 'package:engine_io_shared/packets.dart';
 import 'package:engine_io_shared/transports.dart';
 import 'package:universal_io/io.dart' hide Socket;
 
-import 'package:engine_io_server/src/server/socket.dart';
+import 'package:engine_io_server/src/socket.dart';
 import 'package:engine_io_server/src/transports/transport.dart';
 
 /// Transport used for websocket connections.