+++ packages/server/lib/src/upgrade.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:engine_io_shared/exceptions.dart';
 import 'package:engine_io_shared/transports.dart';
 
-import 'package:engine_io_server/src/server/socket.dart';
+import 'package:engine_io_server/src/socket.dart';
 import 'package:engine_io_server/src/transports/transport.dart';
 
 /// Represents the status of a transport upgrade.