+++ packages/server/test/server/connection_test.dart
@@ -6,8 +6,8 @@ import 'package:engine_io_shared/transports.dart';
 import 'package:test/test.dart';
 import 'package:universal_io/io.dart';
 
-import 'package:engine_io_server/src/server/configuration.dart';
-import 'package:engine_io_server/src/server/server.dart';
+import 'package:engine_io_server/src/configuration.dart';
+import 'package:engine_io_server/src/server.dart';
 import '../matchers.dart';
 import '../shared.dart';