+++ packages/server/test/shared.dart
@@ -5,8 +5,8 @@ import 'package:engine_io_shared/packets.dart';
 import 'package:engine_io_shared/transports.dart';
 import 'package:universal_io/io.dart' hide Socket;
 
-import 'package:engine_io_server/src/server/server.dart';
-import 'package:engine_io_server/src/server/socket.dart';
+import 'package:engine_io_server/src/server.dart';
+import 'package:engine_io_server/src/socket.dart';
 import 'package:engine_io_server/src/transports/polling/polling.dart';
 
 final remoteUrl = Uri.http(InternetAddress.loopbackIPv4.address, '/');