+++ packages/server/CHANGELOG.md
@@ -11,8 +11,9 @@
       contain the session identifier generator and validator functions.
   - [Work in progress] Define the `ClientManager` class to manage references to
     connected clients.
-- [Work in progress] Define the `Socket` class as the super-type of sockets:
+- [Work in progress] Define the `Socket` class to represent clients connected to
+  the server.
   - [Work in progress] Define the `HeartbeatManager` class to manage the
     heartbeat mechanism.
-  - [Work in progress] Define the server/`Socket` class to represent clients
-    connected to the server.
+  - [Work in progress] Define the `UpgradeState` class to manage the state of
+    transport upgrades.